Update user verification and form data in post routes
The method of verifying a user has been adjusted; it now checks for a user object in the session instead of a session itself. Additionally, we now attach "userUid" as a property to form data instead of "username" when creating or updating a post. This offers a more concrete identification method.