Improve post details and comments display
The comments and post details APIs have been extended to include user information, enhancing the data received. The display of comments has also been revised so that the user's profile image is displayed along with the username and comment. Furthermore, the sign-up button in the sign-in page has been temporarily disabled, commented out for future use.