Refactor post model and update post fetch method
This commit represents a refactoring of the post model to remove the 'username' field and replace it with the Wordpress user object. Further, it modifies the post fetch method to include user data. In addition, it improves the date formatted in `view.tsx` of post details by excluding the unwanted time part.