Add sign-out page and improve authentication flow
The changes implemented in this commit include the addition of a new sign-out page and improvement to the existing authentication flow. Now, a user's refresh token is also returned upon signing in. The sign-out page simply triggers the signOut function on render, effectively logging out the user. Also, a new 'verify' route has been introduced to further secure the process.