Remove useModal hook and handle session in Header component
This commit removes the useModal hook as it was no longer needed. Additionally, the way of handling sessions in the Header component has been changed to utilize the next-auth library's useSession hook. This change allows better handling of sessions and their states in a more effective and maintainable way. The session data is now passed directly into the Header component for better encapsulation and ease-of-use, removing unnecessary dependencies.