Add new Button component and move providers to UI
A new Button component was added to the atoms in the UI package. Furthermore, the ChakraUiProvider, ReactQueryProvider, SessionProvider, and ToastProvider were moved from the web app to the UI package. The necessary import statements in dependent files were updated to reflect this change. Also, several new packages were added to the UI package dependencies.