Implement infinite scrolling with intersection observer
Added a new hook 'useInfiniteScroll' that utilizes Intersection Observer API for implementing infinite scrolling functionality in the posts page. Made changes to the posts page to accommodate this hook and to effectively load more posts when the user has scrolled to the end of the page. The 'infinityPosts' query function in 'postsKeys' was also updated to adjust the data fetch based on the page size.