Update Prisma version and implement view count update
This commit introduces an update to Prisma from version 5.13.0 to 5.14.0. Moreover, minor changes have been made to the post view count parameter in the Prisma schema, ensuring that it defaults to 0. Additionally, an axios instance has been implemented to update the view count whenever a post is viewed. Prisma migrations have also been added to the gitignore file.