Add author details query and integration
A new query has been added to fetch author details in our web application. The new file "authorKey.ts" under "queryKeys" includes the definition of this new query. The returned data is then consumed in "view.tsx" for the author page, where it displays the author's username and profile image.