Update user access levels and simplify signIn callback
Changes were made in the Header component to update user access levels for "Write" and "Admin" sections. Now, users can access the "Write" section if their level is 2 or below, and the "Admin" section if their level is 1 or below. In the signIn callback in auth.config.ts, the verification logic was simplified by removing specific conditions for 'google' provider and returning 'true' which enables different verification for other providers.