Add Category interface and associate with tags
Updated the Tag interface in types/tag.ts to include a unique identifier (uid). Implemented a new Category interface that extends Tag and contains an array of associated Tags. Modified queries and routes accordingly to handle the new Category type. Additionally, new functionality was added in the admin view to allow for adding Tags to a Category.