Update writer application querying and logging
Changed the key used in the "createQueryKeys" call from "writer-application" to "writerApplication". Updated the useQuery call in the View component to include a query for writer application details and added a console log for it. Also, improved logging in the writer application route by logging the result of the "findUnique" call.